style: fix all clippy warnings and format code (Phase 25c)
Zero clippy warnings across workspace. Auto-fix unused imports, simplify
boolean expressions, convert indexing loops to iterators, remove dead
assignments. Add #[allow(clippy::field_reassign_with_default)] to 14 test
modules for Java mechanical translation fidelity.

+++ lr2oraja-rust/golden-master/tests/full_pipeline_integration.rs
@@ -45,7 +45,7 @@ fn bms_to_replay_full_pipeline() {
         .collect();
     let replay = ReplayData {
         sha256: Some("test_pipeline".to_string()),
-        mode: model.get_mode().map(|m| m.key() as i32).unwrap_or(0),
+        mode: model.get_mode().map(|m| m.key()).unwrap_or(0),
         keylog: replay_keylog,
         gauge: NORMAL,
         ..Default::default()
@@ -110,7 +110,7 @@ fn full_pipeline_multiple_bms() {
             })
             .collect();
         let replay = ReplayData {
-            mode: model.get_mode().map(|m| m.key() as i32).unwrap_or(0),
+            mode: model.get_mode().map(|m| m.key()).unwrap_or(0),
             keylog: replay_keylog,
             ..Default::default()
         };
